move createWindow function
